+++ R/configuration.R
@@ -62,7 +62,9 @@ config.defaults <- list(
   seq_analysis = list(
     nrepeats = 3,
     stutter.count.ratio_max = 1 / 3,
-    artifact.count.ratio_max = 1 / 3
+    artifact.count.ratio_max = 1 / 3,
+    use_reverse_primers = FALSE,
+    reverse_primer_r1 = TRUE
   ),
   sample_analysis_func = "analyze_sample",
   sample_analysis_opts = list(fraction.min = 0.05),