+++ tests/testthat/test_analyze_seqs.R
@@ -47,6 +47,44 @@ with(test_data, {
     expect_equal(chunk$Artifact, as.integer(NA))
   })
 
+  test_that("analyze_seqs can use reverse primers", {
+    # As is, everything still works if we enable the use of reverse primers from
+    # the locus attributes table.
+    seq_data <- analyze_seqs(seqs1$A, locus_attrs, 3,
+                             use_reverse_primers = TRUE)
+    expect_equal(
+      table(seq_data$MatchingLocus),
+      table(factor(c(rep("A", 14), "B", "B", rep(c("1", "2"), 4)),
+                     levels = c("A", "B", "1", "2"))))
+    # Now, try with a reverse primer that won't match up.
+    locus_attrs_mod <- locus_attrs
+    locus_attrs_mod["A", "ReversePrimer"] <- locus_attrs["B", "ReversePrimer"]
+    seq_data <- analyze_seqs(seqs1$A, locus_attrs_mod, 3,
+                             use_reverse_primers = TRUE)
+    # No more locus A matched since we replaced the reverse primer with B's but
+    # the forward primer still matches A's.
+    expect_equal(
+      table(seq_data$MatchingLocus),
+      table(factor(rep(c("1", "2"), 4), levels = c("A", "B", "1", "2"))))
+  })
+
+  test_that("analyze_seqs can use reverse primers and auto-revcmp", {
+    # If we supply the reverse primers in their orientation on R2, it should
+    # still work as expected so long as we specify reverse_primer_r1 = FALSE.
+    primers <- as.character(
+      Biostrings::reverseComplement(
+        Biostrings::DNAStringSet(locus_attrs[, "ReversePrimer"])))
+    locus_attrs_mod <- locus_attrs
+    locus_attrs_mod[, "ReversePrimer"] <- primers
+    seq_data <- analyze_seqs(seqs1$A, locus_attrs_mod, 3,
+                             use_reverse_primers = TRUE,
+                             reverse_primer_r1 = FALSE)
+    expect_equal(
+      table(seq_data$MatchingLocus),
+      table(factor(c(rep("A", 14), "B", "B", rep(c("1", "2"), 4)),
+                   levels = c("A", "B", "1", "2"))))
+  })
+
   test_that("analyze_seqs checks for motif repeats", {
     seqs <- seqs1$A
     seq_data <- analyze_seqs(seqs, locus_attrs, 3)
@@ -96,7 +134,7 @@ with(test_data, {
   test_that("analyze_seqs works with varied threshold for stutter counts", {
     s <- seqs1$A
     s[nchar(s) %in% c(158, 54)] <- s[nchar(s) == 190][1]
-    seq_data <- analyze_seqs(s, locus_attrs, 3, stutter.count.ratio_max = 1/2)
+    seq_data <- analyze_seqs(s, locus_attrs, 3, stutter.count.ratio_max = 1 / 2)
     chunk <- subset(seq_data, !is.na(Stutter))
     expect_equal(chunk$Count, c(443, 2, 2, 2, 1, 1))
     expect_equal(chunk$Stutter, c(2, 2, 2, 2, 4, 4))
@@ -109,7 +147,7 @@ with(test_data, {
     stutter <- names(sort(table(s), decreasing = TRUE)[3])
     idx <- s == stutter
     s[idx] <- highest
-    substr(s[idx], nchar(stutter), nchar(stutter)) <- "X"
+    substr(s[idx], nchar(stutter), nchar(stutter)) <- "R"
     # Check that the third entry is marked an artifact of the first
     seq_data <- analyze_seqs(s, locus_attrs, 3)
     expect_equal(seq_data$Artifact, c(NA, NA, 1)[1:24])